Moved KbnError base class ⇒ kibana_utils (#45532)
* moved errors base class to kibana_utils

* Fixed tests for ui/errors imports

* fixed test path

* Restored Request Failure resp field

* Restored savedObjectType and savedObjectId

* simplified Error class extension

* updated constructors
