+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns/index_pattern.tsx
@@ -24,14 +24,13 @@ import { EuiButton, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import React from 'react';
 import chrome from 'ui/chrome';
 // @ts-ignore
-import { SavedObjectNotFound, DuplicateField } from 'ui/errors';
-// @ts-ignore
 import { fieldFormats } from 'ui/registry/field_formats';
 // @ts-ignore
 import { expandShorthand } from 'ui/utils/mapping_setup';
 import { toastNotifications } from 'ui/notify';
 import { findObjectByTitle } from 'ui/saved_objects';
 import { SavedObjectsClientContract } from 'src/core/public';
+import { SavedObjectNotFound, DuplicateField } from '../../../../../../plugins/kibana_utils/public';
 
 import { IndexPatternMissingIndices } from '../errors';
 import { Field, FieldList, FieldType } from '../fields';
