+++ src/legacy/ui/public/persisted_state/errors.ts
@@ -17,35 +17,10 @@
  * under the License.
  */
 
-import _ from 'lodash';
-import { ContainerTooSmall } from '../errors';
+import { KbnError } from '../../../../plugins/kibana_utils/public';
 
-/**
- * Common errors shared between constructors
- *
- * @class ErrorHandler
- * @constructor
- */
-export class ErrorHandler {
+export class PersistedStateError extends KbnError {
   constructor() {
-
-  }
-
-  /**
-   * Validates the height and width are > 0
-   * min size must be at least 1 px
-   *
-   * @method validateWidthandHeight
-   * @param width {Number} HTMLElement width
-   * @param height {Number} HTMLElement height
-   * @returns {HTMLElement} HTML div with an error message
-   */
-  validateWidthandHeight(width, height) {
-    const badWidth = _.isNaN(width) || width <= 0;
-    const badHeight = _.isNaN(height) || height <= 0;
-
-    if (badWidth || badHeight) {
-      throw new ContainerTooSmall();
-    }
+    super('Error with the persisted state');
   }
 }
