+++ src/legacy/ui/public/saved_objects/saved_object.js
@@ -31,7 +31,8 @@
 import angular from 'angular';
 import _ from 'lodash';
 
-import { InvalidJSONProperty, SavedObjectNotFound } from '../errors';
+
+import { InvalidJSONProperty, SavedObjectNotFound } from '../../../../plugins/kibana_utils/public';
 import { expandShorthand } from '../utils/mapping_setup';
 
 import { SearchSourceProvider } from '../courier/search_source';
