+++ src/legacy/ui/public/vislib/errors.ts
@@ -1,0 +1,56 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+/* eslint-disable max-classes-per-file */
+
+import { KbnError } from '../../../../plugins/kibana_utils/public';
+
+export class VislibError extends KbnError {
+  constructor(message: string) {
+    super(message);
+  }
+
+  displayToScreen(handler: any) {
+    handler.error(this.message);
+  }
+}
+
+export class InvalidLogScaleValues extends VislibError {
+  constructor() {
+    super('Values less than 1 cannot be displayed on a log scale');
+  }
+}
+
+export class ContainerTooSmall extends VislibError {
+  constructor() {
+    super('This container is too small to render the visualization');
+  }
+}
+
+export class PieContainsAllZeros extends VislibError {
+  constructor() {
+    super('No results displayed because all values equal 0.');
+  }
+}
+
+export class NoResults extends VislibError {
+  constructor() {
+    super('No results found');
+  }
+}
