+++ src/legacy/ui/public/vislib/lib/handler.js
@@ -20,7 +20,7 @@
 import d3 from 'd3';
 import _ from 'lodash';
 import MarkdownIt from 'markdown-it';
-import { NoResults } from '../../errors';
+import { NoResults } from '../errors';
 import { Binder } from '../../binder';
 import { Layout } from './layout/layout';
 import { ChartTitle } from './chart_title';
