+++ src/plugins/kibana_utils/public/index.ts
@@ -19,3 +19,4 @@
 
 export * from './store';
 export * from './parse';
+export * from './errors';
